split up epub-0100
Split epub-test-0100 into:

0100: reflowable content documents
0101: stylesheets
0102: scripting
0103: font embedding & obfuscation
0104: tts

I removed ".. in a reflowable context" from all except the first title,
as it implied that there were a parallel set of FXL tests for things
like fonts, tts, etc; which there aren't.

The reason for splitting this book up is to be able to display more
finely-grained high-level categories and scoring on the EPUBTest
website.

The other option was to modify the EPUBTest website import
configuration to point inside a book, but this was deemed too
complicated to maintain.