Intentionally changed wording
The last commit involved only CSS changes for this test; however, the
epubtest website will acknowledge that a test is new/changed only if
the XHTML changes. So I made a trivial wording change so that this
behavior will be triggered the next time a testsuite gets imported into
the website database.